refactor: move scss modules
