+++ frontend/src/components/RecResource/Card/Activities.scss
