+++ frontend/src/components/RecResource/Card/RecResourceCard.scss
