+++ frontend/src/components/RecResource/PhotoGallery.scss
