+++ frontend/src/components/RecResource/RecResource.scss
