+++ frontend/src/components/Search/Search.scss
