+++ frontend/src/components/layout/Footer.scss
